Make Ollama CVE resolved-version override a feed fallback and co-locate it
